requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
colorama==0.4.6
dj-database-url==2.2.0
Django==4.2.14
django-appconf==1.0.6
django-heroku==0.3.1
django-js-asset==2.2.0
django-quill-editor==0.1.40
django_csp==3.8
gunicorn==22.0.0
libsass==0.23.0
packaging==24.1
psycopg2==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
waitress==3.0.0
whitenoise==6.7.0
